Added Product and OrderDetail functionality

+++ dotnetbackend/MobyLabWebProgramming.Core/Entities/Product.cs
@@ -7,7 +7,7 @@ namespace MobyLabWebProgramming.Core.Entities
         public string Name { get; set; } = default!;
         public string Description { get; set; } = default!;
         public decimal ? Price { get; set; }
-        public decimal? Stock { get; set;}
+        public int? Stock { get; set;}
         public Guid? ProducerId { get; set; } // Foreign Key
         public Producer Producer { get; set; } = default!;
